requirements.txt
pyyaml
streamlit
lightgbm
sklearn
pycaret
altair
pandas